requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=0_gnu
asn1crypto=1.3.0=py37_0
atk=2.32.0=haf93ef1_2
attrs=19.1.0=py_0
bzip2=1.0.8=h516909a_2
ca-certificates=2019.11.28=hecc5488_0
cairo=1.16.0=h18b612c_1001
certifi=2019.3.9=py37_0
cffi=1.14.0=py37hd463f26_0
chardet=3.0.4=py37hc8dfbb8_1006
click=7.1.1=pypi_0
community=1.0.0b1=pypi_0
cryptography=2.5=py37hb7f436b_1
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
decorator=4.4.0=py_0
dill=0.2.9=py37_0
expat=2.2.9=he1b5a44_2
flask=1.1.1=pypi_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.1=he06d7ca_0
gdk-pixbuf=2.32.2=1
gettext=0.19.8.1=hc5be6a0_1002
glib=2.58.3=py37he00f558_1003
gobject-introspection=1.58.2=py37h619baee_1004
graphite2=1.3.13=he1b5a44_1001
graphviz=2.38.0=hf68f40c_1011
gstreamer=1.14.5=h36ae1b5_2
gtk2=2.24.31=h4934977_1002
harfbuzz=2.4.0=h37c48d4_1
html5lib=1.0.1=py_0
icu=58.2=hf484d3e_1000
idna=2.8=py37_1000
intel-openmp=2020.0=166
ipython_genutils=0.2.0=py_1
isodate=0.6.0=py_1
itsdangerous=1.1.0=pypi_0
jinja2=2.11.1=pypi_0
jpeg=9c=h14c3975_1001
jsonschema=3.0.1=py37_0
jupyter_core=4.6.3=py37hc8dfbb8_1
keepalive=0.5=py_1
kiwisolver=1.0.1=py37h6bb024c_1002
klepto=0.1.6=pypi_0
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libffi=3.2.1=he1b5a44_1007
libgcc=7.2.0=h69d50b8_2
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_5
libgomp=9.2.0=h24d8f2e_2
libiconv=1.15=h516909a_1006
liblapack=3.8.0=14_openblas
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hed695b0_1
libprotobuf=3.7.1=h8b12597_0
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc7e4089_6
libtool=2.4.6=h14c3975_1002
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=h13577e0_2
lz4-c=1.8.3=he1b5a44_1001
markupsafe=1.1.1=pypi_0
matplotlib=3.0.3=py37_1
matplotlib-base=3.0.3=py37h5f35d83_1
mkl=2020.0=166
nbformat=4.4.0=py_1
ncurses=6.1=hf484d3e_1002
networkx=2.2=py_1
ninja=1.10.0=hc9558a2_0
numpy=1.16.2=py37h8b7e671_1
olefile=0.46=py_0
openssl=1.0.2u=h516909a_0
pandas=0.24.2=py37hb3f55d8_1
pango=1.40.14=he7ab937_1005
pcre=8.44=he1b5a44_0
pillow=6.0.0=py37he7afcd5_0
pip=20.0.2=py_2
pixman=0.38.0=h516909a_1003
plyfile=0.7=py_0
pox=0.2.5=py_0
protobuf=3.7.1=py37he1b5a44_0
pthread-stubs=0.4=h14c3975_1001
pycparser=2.20=py_0
pygraphviz=1.5=py37h8f50634_1002
pymetis=2019.1.1=pypi_0
pyopenssl=19.0.0=py37_0
pyparsing=2.3.1=py_0
pyqt=5.6.0=py37h13b7fb3_1008
pyrsistent=0.15.2=py37h516909a_0
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.0=hd21baee_1006
python-dateutil=2.8.0=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2018.9=py_0
qt=5.6.2=hce4f676_1013
rdflib=4.2.2=py37_1000
readline=7.0=hf8c457e_1001
requests=2.21.0=py37_1000
retrying=1.3.3=py_2
scikit-learn=0.20.3=py37ha8026db_1
scipy=1.2.1=py37h09a28d5_1
setuptools=46.1.3=py37hc8dfbb8_0
sip=4.18.1=py37hf484d3e_1000
six=1.12.0=py37_1001
sparqlwrapper=1.8.2=py37_1000
sqlite=3.28.0=h8b20d00_0
tensorboardx=1.6=py_0
tk=8.6.10=hed695b0_0
torch-cluster=1.2.4=pypi_0
torch-geometric=1.1.2=pypi_0
torch-scatter=1.1.2=pypi_0
torch-sparse=0.2.4=pypi_0
torch-spline-conv=1.0.6=pypi_0
tornado=6.0.4=py37h8f50634_1
tqdm=4.31.1=py_0
traitlets=4.3.2=py37_1000
urllib3=1.24.1=py37_1000
webencodings=0.5.1=py_1
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py_1
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxpm=3.5.13=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-libxt=1.1.5=h516909a_1003
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h516909a_1002
zlib=1.2.11=h516909a_1006
zstd=1.4.4=h3b9ef0a_2
